Implement creation of a new Author
Receive inputs by form, validate and sanitize them.
Create new Author with form values and encrypt password.
If validateResult function receive any error the form will be redisplay
along with the error and prefilled input fields.
If no errors are registered the new author is save in the database and
redirected to home page.