Update form template view partial
we use this partial for both creation of a new post and to handle an update or delete
of an existing post. When we render the template, if the story is passed in the options,
it means that we are updating or deleting a story. In this case values of inputs are
set to the content of story fields and delete and update buttons are displayed.
If we create a new story values of inputs are set to default as empty strings and
reset and create buttons are displayed